feat: :sparkles: Add parse to allow kd, ks and ksn.
kd, ks, ksn과 같이 object의 특성값을 사용자가 제어할 수 있게 기능 추가

#76
